+++ tests/Dummy3.php
@@ -1,0 +1,42 @@
+<?php
+
+declare(strict_types=1);
+
+namespace voku\tests;
+
+function foo3(int $foo = 0)
+{
+    return new Dummy();
+}
+
+/**
+ * @internal
+ */
+final class Dummy3 implements DummyInterface
+{
+    public $foo;
+
+    /**
+     * @param $foo
+     *
+     * @return mixed
+     */
+    public function lall($foo) {
+        return $foo + 1;
+    }
+
+    /**
+     * @param \phpDocumentor\Reflection\DocBlock\Tags\BaseTag $parsedParamTag
+     *
+     * @return array
+     *
+     * @psalm-return array{parsedParamTagStr: string, variableName: null[]|string}
+     */
+    public function withComplexReturnArray(\phpDocumentor\Reflection\DocBlock\Tags\BaseTag $parsedParamTag)
+    {
+        return [
+            'parsedParamTagStr' => 'foo',
+            'variableName'      => [null],
+        ];
+    }
+}